orchestrator/pyproject.toml
[project]
name = "orchestrator"
version = "0.5.0"
description = "Buttercup orchestrator"
readme = "README.md"
authors = [{ name = "Trail of Bits", email = "opensource@trailofbits.com" }]
requires-python = ">=3.12,<3.13"
dependencies = [
    "argon2-cffi>=21.0.0",
    "common",
    "fastapi>=0.115.6",
    "pydantic>=2.10.5",
    "pydantic-settings>=2.7.1",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.1",
    "requests>=2.32.3",
    "requests-file>=2.1.0",
    "rich>=13.9.4",
    "sqlalchemy>=2.0.0",
    "typing-extensions>=4.12.2",
    "urllib3>=2.3.0",
    "uvicorn>=0.34.0",
    "nats-py>=2.8.0",
]

[project.scripts]
buttercup-task-server = "buttercup.orchestrator.task_server.__cli__:main"
buttercup-task-downloader = "buttercup.orchestrator.downloader.__cli__:main"
buttercup-scheduler = "buttercup.orchestrator.scheduler.__cli__:main"
buttercup-pov-reproducer = "buttercup.orchestrator.pov_reproducer.__cli__:main"
buttercup-auth-tool = "buttercup.orchestrator.task_server.auth_tool:main"
buttercup-scratch-cleaner = "buttercup.orchestrator.scratch_cleaner.__cli__:main"
buttercup-ui = "buttercup.orchestrator.ui.__cli__:main"


[project.optional-dependencies]
tests = [
    "flake8>=7.1.1",
    "mypy>=1.15.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.6.1",
    "ruff>=0.9.2",
    "types-python-dateutil>=2.9.0.20241206",
    "types-requests>=2.32.0",
    "fastapi[standard]>=0.115.6",
    "responses>=0.25.6",
    "httpx>=0.28.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/buttercup"]

[tool.uv.sources]
common = { path = "../common", editable = true }

[tool.ruff]
line-length = 120
exclude = [
    "src/buttercup/orchestrator/competition_api_client",
    "test/test_types_*.py",
    "test/test_vulnerability_api.py",
    "test/test_patch_api.py",
    "test/test_sarif_api.py",
    "src/buttercup/orchestrator/task_server/models/",
]

[tool.mypy]
plugins = ["pydantic.mypy"]
mypy_path = "src"
packages = "buttercup"
allow_redefinition = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
sqlite_cache = true
strict_equality = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true
exclude = ['competition_api_client', 'ui/competition_api/models']
